[ FIX ] 긴급 수정 - JpaRepository 기본키 타입 수정

+++ src/main/java/com/itwillbs/factron/repository/approval/ApprovalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApprovalRepository extends JpaRepository<Approval, Integer> {
+public interface ApprovalRepository extends JpaRepository<Approval, Long> {
 }
